Move around things, remove other things
