+++ doc.go
@@ -1,5 +1,9 @@
 /*Package libs is a collection of Go libraries for building Altid filesystems
+service
+===========
 
-Refer to altid.github.io for an overview
+Package libs/service provides the core for Altid services.
+Included are the network listener and the file-backed state for services. A service can interact through the use of the various FileWriters; as well as register a callback function to handle control messages and user input. 
+
 */
 package libs
