--- server/README.md
@@ -1,10 +1,0 @@
-# server
-Altid server library
-
-[![GoDoc](https://godoc.org/github.com/golang/gddo?status.svg)](https://godoc.org/github.com/altid/server) ![Tests](https://github.com/altid/server/workflows/Tests/badge.svg) [![Go Report Card](https://goreportcard.com/badge/github.com/altid/server)](https://goreportcard.com/report/github.com/altid/server) [![License](http://img.shields.io/:license-mit-blue.svg)](http://doge.mit-license.org)
-
-## Usage
-
-These are compilation-time plugins for use in Altid services
-
-TODO: Example usage
