--- server/client/doc.go
@@ -1,6 +1,0 @@
-/*Package client provides management of clients for a server
-
-go get github.com/altid/server/client
-
-*/
-package client
