--- server/doc.go
@@ -1,6 +1,0 @@
-/*Package server is a library used to provide server capabilities to Altid services 
-
-	go get github.com/altid/server
-
- */
-package server
