--- server/go.mod
@@ -1,12 +1,0 @@
-module github.com/altid/server
-
-go 1.14
-
-require (
-	github.com/altid/libs v0.1.6-patched
-	github.com/go9p/styx v0.0.0-20200314201655-4aa21a63d347
-	github.com/google/uuid v1.1.1
-	github.com/grandcat/zeroconf v1.0.0
-	golang.org/x/crypto v0.0.0-20191011191535-87dc89f01550
-	golang.org/x/sys v0.0.0-20200409092240-59c9f1ba88fa // indirect
-)
