+++ service/TODO
@@ -1,0 +1,10 @@
+Remove need for input:foo:message, just have input type and no weird fucking parsing
+Remove Mock/Default, just use in-memory and toggle off logfiles for testing
+Encrypt cache, etc?
+Cleanup just cleans up our cache if we have to use it 
+Listen or Start to run all the commands
+ReadWriters instead of Readers and Writers
+Create an input per client connect
+ - Instead of having an input on the service like we have, attach it to the buffer and pass the handler over to it
+ - This obviates the range and fuckiness around that
+
