+++ service/command_test.go
@@ -1,4 +1,4 @@
-package fs
+package service 
 
 import (
 	"testing"
