+++ service/doc.go
@@ -1,0 +1,8 @@
+/*
+Package service is a library which allows you to write well constructed Altid services
+
+Overview
+
+service aims to present a way to write canonical services, which behave correctly in all instances.
+*/
+package service
