+++ service/internal/command/lexer.go
