+++ service/internal/reader/cmd.go
