+++ service/internal/reader/cmd_test.go
